Pfem interface separated from fsi + some cleaning
